fix(task): 修复了produce模块中循环引用，导致produce子模块失效的问题
